feat: rearrange result page component
