+++ components/test/results/CopyLinkBtn.tsx
@@ -108,8 +108,7 @@ const CopyLinkWrapper = styled.button`
     cursor: pointer;
   }
   ${media.mobile} {
-    margin-top: 1.222rem;
-    margin-left: 4.6rem;
+    margin-top: 0rem;
     width: 11.8rem;
     height: 3.274rem;
     font-size: 1.2rem;
