+++ components/test/results/MatchingColor1.tsx
@@ -35,17 +35,9 @@ const ColorWrap = styled.button`
   font-weight: 500;
 
   ${media.mobile} {
-    font-size: 1.4rem;
-  }
-`;
-
-const Text = styled.span`
-  text-align: center;
-  color: ${({ theme }) => theme.colors.white};
-  font-size: 1.8rem;
-  font-weight: 500;
-
-  ${media.mobile} {
+    margin-right: 0.35rem;
+    margin-left: 0.35rem;
+    /* padding: 0; */
     font-size: 1.4rem;
   }
 `;
